+++ src/types.ts
@@ -71,9 +71,7 @@ export enum StorageKey {
     SERVER_SIDE_FILTERS = 'ServerFilters',
     TIMELINE = 'Timeline',  // The entire time line (home timeline + trending toots etc.)
     TRENDING_TAG_TOOTS = 'TrendingTagToots',
-    TOOT_TESTER = 'TootTester',
     USER = 'FedialgoUser',
-    USER_DATA = 'UserData',  // Only used for mutex for now
     WEIGHTS = 'Weights',
 };